requirements.txt
pytesseract
pillow
opencv-python
numpy
google-cloud-vision
cachetools
twilio
flask
